+++ openyich-framework/src/main/java/com/openyich/framework/boot/autoconfigure/OpenYichAutoConfiguration.java
@@ -5,6 +5,7 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
+import com.openyich.framework.boot.config.AspectJConfiguration;
 import com.openyich.framework.boot.config.AsyncConfiguration;
 import com.openyich.framework.boot.config.DateTimeFormatConfiguration;
 import com.openyich.framework.boot.config.ExceptionTranslator;
@@ -28,6 +29,7 @@ import com.openyich.framework.boot.web.DefaultResponseWordsAware;
 @ConditionalOnWebApplication
 @EnableConfigurationProperties({OpenYichProperties.class})
 @Import({
+  AspectJConfiguration.class,
   AsyncConfiguration.class, 
   DateTimeFormatConfiguration.class,
   DefaultHttpHeadersAware.class,