+++ openyich-framework/src/main/java/com/openyich/framework/boot/web/DefaultHttpHeadersAware.java
@@ -1,5 +1,7 @@
 package com.openyich.framework.boot.web;
 
+import java.util.UUID;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -9,6 +11,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
 
 import com.openyich.framework.boot.aware.HttpHeadersAware;
+import com.openyich.framework.boot.utils.ThreadLocalUtils;
+import com.openyich.framework.boot.utils.ThreadLocalUtils.RequestHeader;
 
 @Configuration
 @ConditionalOnWebApplication
@@ -18,6 +22,13 @@ public class DefaultHttpHeadersAware implements HttpHeadersAware {
   @Override
   public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler,
       HttpHeaders httpHeaders) {
+    RequestHeader requestHeader = new RequestHeader();
+    requestHeader.setHttpHeaders(httpHeaders);
+    requestHeader.setRequest(request);
+    requestHeader.setResponse(response);
+    requestHeader.setSessionId(request.getRequestedSessionId());
+    requestHeader.setRequestId(UUID.randomUUID().toString());
+    ThreadLocalUtils.set(requestHeader);
     return true;
   }